+ added JSON error reporting for ES when issuing invalid queries
+ ResultsManager modified so that ES queries are issued whenever the tql changes, and fixed bugs with not issuing
queries when they should be

+ Using "clarinet" json parser to ensure proper line, column, and position for error messages

+ May return multiple parser errors, ordered by position within the query. Currently only the first error is reported
because of restrictions in the legacy query API

+ Changed the query route API for ES slightly to allow a string to be passed instead of an object as the query body.
This is intended to be used for detecting JSON errors, but may be used to pass an unparsed query string as well. The API
spec allows and allows "any" type to be passed as the query body, so this change only affects the API when querying an
ES database via the query route.