Remove use of "export namespace".
We should just be using TS-style external modules -- the linter
complains if either the `namespace` or the `module` keyword is used.

`module Foo {` and `namespace Foo {` are equivalent and the use of
either is discouraged

From the handbook:
```
To reiterate why you shouldn't try to namespace your module contents,
the general idea of namespacing is to provide logical grouping of
constructs and to prevent name collisions. Because the module file
itself is already a logical grouping, and its top-level name is defined
by the code that imports it, it's unnecessary to use an additional
module layer for exported objects.
```