More robust input check in cards-to-elastic.
Earlier, we would just check if a block's value started with @ and
assume that the value is an "input" and return it directly. This
worked for primitive blocks/types but not for blocks that return objects
or rely on recursion to generate its TQL.

This makes the input check slightly more robust (although not ideal) in
that we call the translation function for a block and check if the
returned value is undefined or NaN and an input to treat the value as an
input.