Initialize a variant's deployed name when it's duplicated.
When a variant is duplicated, we did not initialize its deployed name---the old variant's deployed name
was simply copied over. This results in an issue in the analytics tab (where the deployed name is used
to fetch a variant's metrics data) where a variant showed the data for another variant that it was duplicated
from.